requirements.txt
fastapi
uvicorn
requests
faiss-cpu
sentence-transformers
python-dotenv
pandas
google-generativeai